requirements.txt
discord.py==2.2.3
fastapi==0.95
python-dotenv==1.0.0
uvicorn==0.22.0
pydantic~=1.10.7
aiohttp~=3.8.4
loguru==0.7.0
aiofiles==23.1.0
python-multipart==0.0.6
deepl==1.15.0
APScheduler==3.10.1
retrying==1.3.4
Pillow==8.4.0
requests_toolbelt==1.0.0